Add tests for maps and collection with groups and refactor assertions
The added tests check that the configuration validation works for maps
and collections on method level, on the keys and nested collections.
Three of the five new validation problems are missed in the current
implementation.

The reason for the custom assertion is that the old approach was not as
informative as it already failed at the size check. The new approach
also reduces the danger that the size assertion and the number of
individual contains checks deviate.
This is already implemented in test frameworks such as AssertJ, but
those are not available here and thus a custom solution is required for
now, unless we can add a test dependency.