Provide some shortcuts for simple config properties
This avoids a bunch of allocations, including the allocation of an
ObjectCreator for each property.
It was kept simple on purpose as a base for discussions.

Note that, interestingly enough, my startup actually got a bit slower
with this patch which makes little sense... I thought it might be
interesting to discuss this line of thoughts anyway.